+++ packages/amplication-client/src/Resource/useProjectConfigSettingsHook.tsx
@@ -1,8 +1,9 @@
 import * as models from "../models";
 import { useCallback } from "react";
+import { gql } from "@apollo/client";
+import { useTracking } from "../util/analytics";
 
 interface ProjectSettingsHookParams {
-  trackEvent: (event: { eventName: string; [key: string]: any }) => void;
   resourceId: string;
   updateResourceSettings: any;
 }
@@ -17,10 +18,11 @@ const PROJECT_CONFIG_FORM_SCHEMA = {
 };
 
 const useProjectConfigSettingsHook = ({
-  trackEvent,
   resourceId,
   updateResourceSettings,
 }: ProjectSettingsHookParams) => {
+  const { trackEvent } = useTracking();
+
   const handleSubmit = useCallback(
     (data: models.ProjectConfigurationSettings) => {
       const { baseDirectory } = data;
@@ -46,3 +48,27 @@ const useProjectConfigSettingsHook = ({
 };
 
 export default useProjectConfigSettingsHook;
+
+export const UPDATE_PROJECT_CONFIG_SETTINGS = gql`
+  mutation updateProjectConfigurationsSettings(
+    $data: ProjectConfigurationSettingsUpdateInput!
+    $resourceId: String!
+  ) {
+    updateProjectConfigurationSettings(
+      data: $data
+      where: { id: $resourceId }
+    ) {
+      id
+      baseDirectory
+    }
+  }
+`;
+
+export const GET_PROJECT_CONFIG_SETTINGS = gql`
+  query projectConfigurationSettings($id: String!) {
+    projectConfigurationSettings(where: { id: $id }) {
+      id
+      baseDirectory
+    }
+  }
+`;
